refactor: move getLanguageCodes and textifyReadableStream to util.ts
- Move getLanguageCodes from globalApi.svelte.ts to util.ts
- Add uiLanguage parameter to getLanguageCodes for explicit dependency
- Move textifyReadableStream from globalApi.svelte.ts to util.ts
- Update imports in PlaygroundTranslation, PlaygroundSubtitle, and request handlers
