+++ src/ts/process/request/openAI.ts
@@ -4,7 +4,7 @@ import { getDatabase } from "src/ts/storage/database.svelte"
 import { LLMFlags, LLMFormat } from "src/ts/model/modellist"
 import { strongBan, tokenizeNum } from "src/ts/tokenizer"
 import { getFreeOpenRouterModel } from "src/ts/model/openrouter"
-import { textifyReadableStream } from "src/ts/globalApi.svelte"
+import { textifyReadableStream } from "src/ts/util"
 import { globalFetch, fetchNative, addFetchLog } from "src/ts/fetch"
 import { isTauri, isNodeServer, isCapacitor } from "src/ts/platform"
 import type { OpenAIChatFull } from "../index.svelte"
